requirements.txt
Requests==2.31.0
requests_html==0.10.0
pyyaml==6.0.1
